backend: Improve Save()
As mentioned in issue [#1560](https://github.com/restic/restic/pull/1560#issuecomment-364689346)
this changes the signature for `backend.Save()`. It now also takes a
function parameter which yields an `io.Reader` (and an error), so that
the backend implementations or our `RetryBackend` middleware can request
a new, fresh reader from the caller to retry an upload operation.

In addition, this adds the `length` parameter, which is used to tell the
backend how many bytes are there to be saved. This removes several ugly
hacks we had to do to pull the size back out of the `io.Reader` passed
to `Save()` before. In the `s3` and `rest` backend this is actively
used.